refactor(Convo): consolidate logic to navigateWithLastTools from useNavigateToConvo
